requirements.txt
asgiref==3.7.2
distlib==0.3.7
Django==5.0.1
django-crispy-forms==2.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
filelock==3.12.4
joblib==1.3.2
numpy==1.26.0
Pillow==10.1.0
platformdirs==3.10.0
pytz==2019.3
scikit-learn==1.3.0
scipy==1.11.2
six==1.13.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tzdata==2023.3
virtualenv==20.24.5
xgboost==1.7.6
pandas==2.1.3
matplotlib==3.8.2
tensorflow==2.15.0
torchvision==0.15.2
